pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doc2md"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "typer",
    "pymupdf",
    "pyyaml",
    "python-magic",
    "pydantic>=2",
]

[project.scripts]
doc2md = "doc2md.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src/doc2md"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
strict_optional = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["fitz", "yaml"]
ignore_missing_imports = true
